vmdk-to-vhd : move all VHD code to vhd-lib
