+++ @xen-orchestra/vhd-lib/src/index.js
@@ -5,13 +5,4 @@ export {
   readVhdMetadata,
 } from './vhd-merge'
 
-export {
-  Vhd,
-  fuHeader,
-  fuFooter,
-  checksumStruct,
-  VHD_SECTOR_SIZE,
-  HARD_DISK_TYPE_FIXED,
-  HARD_DISK_TYPE_DYNAMIC,
-  computeGeometryForSize,
-} from './vhd'
+export { Vhd, ReadableRawVHDStream, computeGeometryForSize } from './vhd'
