+++ @xen-orchestra/vhd-lib/src/vhd-merge.integ.spec.js
@@ -5,9 +5,9 @@ import fs from 'fs-extra'
 import getStream from 'get-stream'
 import rimraf from 'rimraf'
 import tmp from 'tmp'
+import { getHandler } from '@xen-orchestra/fs'
 import { randomBytes } from 'crypto'
 import { fromCallback as pFromCallback, fromEvent } from 'promise-toolbox'
-import { getHandler } from '@xen-orchestra/fs'
 
 import { VHD_SECTOR_SIZE, Vhd } from './vhd'
 import vhdMerge, { chainVhd, createReadStream } from './vhd-merge'
