+++ packages/xo-vmdk-to-vhd/package.json
@@ -46,7 +46,9 @@
     "@babel/preset-flow": "7.0.0-beta.42",
     "babel-plugin-lodash": "^3.3.2",
     "cross-env": "^5.1.3",
+    "event-to-promise": "^0.8.0",
     "execa": "^0.10.0",
+    "fs-extra": "^5.0.0",
     "index-modules": "^0.3.0",
     "rimraf": "^2.6.2"
   },
